Refactor for improved code reuse and performance, added restart function and includes changes to the api
Change to Animatr api, the css3 animation properties need to be defined in the same way as the w3c spec, i.e. the Animatr previously accepted an integer/float as a value e.g. 1 but now it must have confirm to w3c so it is now '1s' note the addition of the 's' and its now a string. (same for the delay property)